New cross validation alignment algo
The sensor datapoints are now aligned by finding all of the timestamps
that matchup and discarding the rest instead of just finding the first
pair that match and assuming the rest line up.

Tests would fail before this because CHRE would often give sensor data
at an increased rate as compared to the AP because CHRE would not
resample the data.

Fixes: 174659971
Bug: 175795665
Test: All cross validation tests pass
Change-Id: Iee37a349020f12b6651543011cb74cd9ebcfd01c