Assert on invalid size in dynamic vector operations
This CL adds asserts at various places that 'mSize' of the dynamic
vector is accessed/changed, to help check for possible heap corruption.
Since there are already asserts on the index being less than size,
this CL is expected to be temporary, and reverted after the underlying
issue has been identified.

Bug: 178963854
Test: Compile
Change-Id: I345452b7cc66c5e1ce5f220588d6df5783039dee