Do not send another buffered log if first one fails
If the daemon is disconnected from CHRE for long time, and infinite loop
occurs because the onLogsSentToHost callback will not clear the
secondary buffer but will trigger another attempt to send the logs which
will start the loop.

Instead just have the flag to send another log through depend on if the
first log send passed.

Fixes: 184839292
Test: Observe logs and do not see an infinite loop if the daemon
disconnects

Change-Id: I1db52e5eb19ce2d1007ff16de42e6d0d6f70bde9