Use unstable nanoapp ID in place of version string
All nanoapps will now use a symbol that is in a different section of
memory that can be stripped out before hashing the binary. This unstable
id will potentially change without any changes to the underlying
functionality of the nanoapp because a new unstable ID will be created
for each build of the nanoapp regardless of code changes.

Also, the NANOAPP_VERSION_STRING flag will be fully replaced for all
platforms with the NANOAPP_UNSTABLE_ID.

Bug: 166270600
Test: run load_app.sh with the activity nanoapp and observe
unstable_nanoapp_id

Change-Id: I53ff54ea3792592d59b79908d99dc129f864e742