Infer the section alignment size from the header.
Our nanoapp loader currently assumes that the nanoapp elf binaries
have a fixed 4k alignment. This can be wasteful for small nanoapps
and can break the implementation for binaries with possibly larger
alignments. We can infer the alignment for the section from the
section header, and allocate memory based off of this.

Bug: 194846124
Test: Load changes, ensure nanoapps are loaded as expected.
Change-Id: I9ebc1fddbc65fe730e334d9499e4d2f27af85500