Fix output path and options warning in nanopb.mk
Instead of depending on the (undocumented!) input variable
NANOPB_PROTO_PATH to build the path for protoc to place generated .c/.h
files, extract this from the target rule for the .c/.h file itself,
which avoids a build failure when the two don't match.

Also, only supply the --options-file parameter if NANOPB_OPTIONS is not
empty, which fixes a warning for "Options file not found:" in this
situation.

As a bonus, also properly document all of the variables that nanopb.mk
can accept.

Bug: 218381518
Test: test CHRE & nanoapp compilation
Change-Id: Iac3b33ab73063a86a42fb4bad721abe86d4b439b