Fix build template expansion
In the Makefile, the nanoapp build template has incorrect number of
dollar signs at several places, leading to incorrect expansion.  This
causes the object files to be generated at a path without the target
name in it.  As a result, when building more than one targets
concurrently, we have a race condition that can result in the corruption
of object files sometimes.

Bug: 207703544
Test: Repeated runs of build_chre.sh do not fail.
Test: After a build, check object files are now generated in per-target
      locations.
Change-Id: I0abb14906676372a7f4c5ad7eab44498ec175545